+++ crates/wasm/src/lib.rs
@@ -35,13 +35,23 @@ impl crate::wasmer_pack::Package for Package {
         Ok(Handle::new(Package(pkg)))
     }
 
-    fn generate_javascript(&self) -> Result<Vec<wasmer_pack::File>, wasmer_pack::Error> {
-        let files = original::generate_javascript(&self.0)?;
+    fn generate_javascript(
+        &self,
+        options: wasmer_pack::BindingsOptions,
+    ) -> Result<Vec<wasmer_pack::File>, wasmer_pack::Error> {
+        let files = original::generate_javascript(
+            &self.0,
+            original::BindingsOptions { name: options.name },
+        )?;
         Ok(unwrap_files(files))
     }
 
-    fn generate_python(&self) -> Result<Vec<wasmer_pack::File>, wasmer_pack::Error> {
-        let files = original::generate_python(&self.0)?;
+    fn generate_python(
+        &self,
+        options: wasmer_pack::BindingsOptions,
+    ) -> Result<Vec<wasmer_pack::File>, wasmer_pack::Error> {
+        let files =
+            original::generate_python(&self.0, original::BindingsOptions { name: options.name })?;
         Ok(unwrap_files(files))
     }
 }