+++ crates/wasmer-pack/src/py/mod.rs
@@ -8,7 +8,7 @@ use wai_bindgen_gen_core::Generator;
 use wai_bindgen_gen_wasmer_py::WasmerPy;
 
 use crate::{
-    types::{Interface, Package},
+    types::{BindingsOptions, Interface, Package},
     Files, Metadata, Module, SourceFile,
 };
 
@@ -36,9 +36,15 @@ static TEMPLATES: Lazy<Environment> = Lazy::new(|| {
 });
 
 /// Generate Python bindings.
-pub fn generate_python(package: &Package) -> Result<Files, Error> {
+pub fn generate_python(package: &Package, options: BindingsOptions) -> Result<Files, Error> {
     let metadata = package.metadata();
-    let package_name = metadata.package_name.python_name();
+
+    // make sure the name is in snake-case
+    let package_name = if let Some(name) = options.name {
+        name.to_snake_case()
+    } else {
+        metadata.package_name.name().to_string().to_snake_case()
+    };
 
     let mut files = Files::new();
 
@@ -167,6 +173,7 @@ impl From<Interface> for InterfaceContext {
 
 #[derive(Debug, serde::Serialize)]
 struct CommandContext {
+    name: String,
     ident: String,
     module_filename: String,
     #[serde(skip)]
@@ -175,11 +182,12 @@ struct CommandContext {
 
 impl From<crate::Command> for CommandContext {
     fn from(cmd: crate::Command) -> CommandContext {
-        let ident = cmd.name.replace('-', "_");
-        let module_filename = format!("{ident}.wasm");
+        let ident = cmd.name.to_snake_case();
+        let module_filename = Path::new(&ident).with_extension("wasm");
         CommandContext {
+            name: cmd.name.clone(),
             ident,
-            module_filename,
+            module_filename: module_filename.display().to_string(),
             wasm: cmd.wasm,
         }
     }
@@ -393,7 +401,7 @@ mod tests {
         }];
         let package = Package::new(metadata, libraries, commands);
 
-        let files = generate_python(&package).unwrap();
+        let files = generate_python(&package, BindingsOptions::default()).unwrap();
 
         let actual_files: BTreeSet<_> = files.iter().map(|(p, _)| p).collect();
         assert_eq!(actual_files, expected);